+++ src/GLViewer/GLViewer_ViewPort2d.cxx
@@ -1367,7 +1367,7 @@ void GLViewer_ViewPort2d::onMaybeTip( QPoint thePoint, QString& theText, QFont& 
       int str_size = aList.first().length();
       for( int i = 1, size = aList.count(); i < size; i++ )
       {
-        if( str_size < aList[i].length() )
+        if ( str_size < (int)aList[i].length() )
         {
           index = i;
           str_size = aList[i].length();
